% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fucking_functions.R
\name{foaas_deraadt}
\alias{foaas_deraadt}
\title{{name} you are being the usual slimy hypocritical asshole... You may have had value ten years ago, but people will see that you don't anymore.}
\usage{
foaas_deraadt(name, from)
}
\description{
{name} you are being the usual slimy hypocritical asshole... You may have had value ten years ago, but people will see that you don't anymore.
}

